requirements.txt
pandas
numpy
seaborn
streamlit
matplotlib
pyyaml